backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.20
google-generativeai==0.8.3
PyPDF2==3.0.1
python-docx==1.1.2
pydantic==2.10.3
python-dotenv==1.0.1
